Fix formatting of adjacent long words. (#9)
If two words that are exceeding the line length follow each other, the
algorithm previously pulled them on a single line because this way the
penalty was paid only once.

This patch fixes this behavior by not adding the penalty if the line is
exceeded due to a long word, i.e. when it is unavoidable.